Reorganization of the project. Added sale models.
