+++ src/domain/Comic.java
@@ -1,4 +1,4 @@
-package models;
+package domain;
 
 /**
  * Модель комикса
