+++ src/domain/sale/Customer.java
@@ -1,0 +1,10 @@
+package domain.sale;
+
+public class Customer {
+
+    private String name;
+
+    public Customer(String name) {
+        this.name = name;
+    }
+}
